Adds auto jump to next page button
Implements a button that automatically navigates to the next page after a countdown.

This button appears when there is no data on the current page,
but there are more pages available.

It also handles the case where the last page has no data,
prompting the user to jump to the previous page.